Safari-611.1.10/LayoutTests/crypto/subtle/hmac-generate-key-sha384-expected.txt

Test generating a HMAC key with sha-384.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Generating a key...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'HMAC'
PASS key.algorithm.length is 1024
PASS key.algorithm.hash.name is 'SHA-384'
PASS key.usages is ["sign", "verify"]
PASS successfullyParsed is true

TEST COMPLETE
